refactor: UI refactor
