+++ packages/alice/lib/ui/call_details/page/alice_call_details_page.dart
@@ -2,10 +2,10 @@ import 'package:alice/core/alice_core.dart';
 import 'package:alice/helper/alice_save_helper.dart';
 import 'package:alice/model/alice_http_call.dart';
 import 'package:alice/ui/call_details/model/alice_call_details_tab.dart';
-import 'package:alice/ui/call_details/widget/alice_call_error_widget.dart';
+import 'package:alice/ui/call_details/widget/alice_call_error_screen.dart';
 import 'package:alice/ui/call_details/widget/alice_call_overview_screen.dart';
-import 'package:alice/ui/call_details/widget/alice_call_request_widget.dart';
-import 'package:alice/ui/call_details/widget/alice_call_response_widget.dart';
+import 'package:alice/ui/call_details/widget/alice_call_request_screen.dart';
+import 'package:alice/ui/call_details/widget/alice_call_response_screen.dart';
 import 'package:alice/ui/common/alice_page.dart';
 import 'package:alice/utils/alice_theme.dart';
 import 'package:collection/collection.dart' show IterableExtension;
