+++ packages/alice/lib/ui/calls_list/widget/alice_logs_screen.dart
@@ -4,8 +4,8 @@ import 'package:alice/ui/calls_list/widget/alice_log_list_widget.dart';
 import 'package:alice/ui/calls_list/widget/alice_raw_log_list_widger.dart';
 import 'package:flutter/material.dart';
 
-class AliceLogsWidget extends StatelessWidget {
-  const AliceLogsWidget({
+class AliceLogsScreen extends StatelessWidget {
+  const AliceLogsScreen({
     super.key,
     required this.scrollController,
     this.aliceLogger,
